+++ CHANGELOG.md
@@ -25,7 +25,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ### Changed
 - process_attach:  Add ability to capture multiple attachments in an email.
-- process_message:  Refactored function and added ability to process emails based on the email's From address.
+- process_message:  Refactored function and added ability to process emails based on the emails From address.
 - config/rabbitmq.py.TEMPLATE:  Added heartbeat, host_list, err_addr_queue and queue_dict entries.
 - Documentation updates.